third_party/blink/web_tests/external/wpt/css/css-content/parsing/content-invalid-expected.txt

This is a testharness.js-based test.
Found 35 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['content'] = "counters(counter-name) / \\"alt text\\" attr(foo) \\"bar\\"" should not set the property value
  assert_equals: expected "" but got "counters(counter-name) / \\"alt text\\" attr(foo) \\"bar\\""
[FAIL] e.style['content'] = "counters(counter-name) / attr(foo)" should not set the property value
  assert_equals: expected "" but got "counters(counter-name) / attr(foo)"
[FAIL] e.style['content'] = "counter() / \\"alt text\\" attr(foo) \\"bar\\"" should not set the property value
  assert_equals: expected "" but got "counter() / \\"alt text\\" attr(foo) \\"bar\\""
[FAIL] e.style['content'] = "counter() / attr(foo)" should not set the property value
  assert_equals: expected "" but got "counter() / attr(foo)"
[FAIL] e.style['content'] = "open-quote / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "open-quote / attr(foo) open-quote"
[FAIL] e.style['content'] = "close-quote / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "close-quote / attr(foo) open-quote"
[FAIL] e.style['content'] = "no-open-quote / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "no-open-quote / attr(foo) open-quote"
[FAIL] e.style['content'] = "no-close-quote / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "no-close-quote / attr(foo) open-quote"
[FAIL] e.style['content'] = "attr(alt) / url(\\"https://www.example.com/picture.svg\\")" should not set the property value
  assert_equals: expected "" but got "attr(alt) / url(\\"https://www.example.com/picture.svg\\")"
[FAIL] e.style['content'] = "attr(alt) / no-open-quote" should not set the property value
  assert_equals: expected "" but got "attr(alt) / no-open-quote"
[FAIL] e.style['content'] = "attr(alt) / no-close-quote" should not set the property value
  assert_equals: expected "" but got "attr(alt) / no-close-quote"
[FAIL] e.style['content'] = "attr(alt) / \\"hi\\" no-close-quote" should not set the property value
  assert_equals: expected "" but got "attr(alt) / \\"hi\\" no-close-quote"
[FAIL] e.style['content'] = "attr(alt) / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "attr(alt) / attr(foo) open-quote"
[FAIL] e.style['content'] = "attr(data-foo) / url(\\"https://www.example.com/picture.svg\\")" should not set the property value
  assert_equals: expected "" but got "attr(data-foo) / url(\\"https://www.example.com/picture.svg\\")"
[FAIL] e.style['content'] = "attr(data-foo) / no-open-quote" should not set the property value
  assert_equals: expected "" but got "attr(data-foo) / no-open-quote"
[FAIL] e.style['content'] = "attr(data-foo) / no-close-quote" should not set the property value
  assert_equals: expected "" but got "attr(data-foo) / no-close-quote"
[FAIL] e.style['content'] = "attr(data-foo) / \\"hi\\" no-close-quote" should not set the property value
  assert_equals: expected "" but got "attr(data-foo) / \\"hi\\" no-close-quote"
[FAIL] e.style['content'] = "attr(data-foo) / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "attr(data-foo) / attr(foo) open-quote"
[FAIL] e.style['content'] = "counter(counter-name) / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "counter(counter-name) / attr(foo) open-quote"
[FAIL] e.style['content'] = "counter(counter-name, counter-style) / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "counter(counter-name, counter-style) / attr(foo) open-quote"
[FAIL] e.style['content'] = "counters(counter-name, \\".\\") / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "counters(counter-name, \\".\\") / attr(foo) open-quote"
[FAIL] e.style['content'] = "counters(counter-name, \\".\\", counter-style) / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "counters(counter-name, \\".\\", counter-style) / attr(foo) open-quote"
[FAIL] e.style['content'] = "url(\\"https://www.example.com/picture.svg\\") / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "url(\\"https://www.example.com/picture.svg\\") / attr(foo) open-quote"
[FAIL] e.style['content'] = "\\"hello\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" / attr(foo) open-quote"
[FAIL] e.style['content'] = "\\"hello\\" \\"world\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" \\"world\\" / attr(foo) open-quote"
[FAIL] e.style['content'] = "\\"hello\\" attr(alt) \\"world\\" / url(\\"https://www.example.com/picture.svg\\")" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" attr(alt) \\"world\\" / url(\\"https://www.example.com/picture.svg\\")"
[FAIL] e.style['content'] = "\\"hello\\" attr(alt) \\"world\\" / no-open-quote" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" attr(alt) \\"world\\" / no-open-quote"
[FAIL] e.style['content'] = "\\"hello\\" attr(alt) \\"world\\" / no-close-quote" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" attr(alt) \\"world\\" / no-close-quote"
[FAIL] e.style['content'] = "\\"hello\\" attr(alt) \\"world\\" / \\"hi\\" no-close-quote" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" attr(alt) \\"world\\" / \\"hi\\" no-close-quote"
[FAIL] e.style['content'] = "\\"hello\\" attr(alt) \\"world\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "\\"hello\\" attr(alt) \\"world\\" / attr(foo) open-quote"
[FAIL] e.style['content'] = "counter(counter-name) \\"potato\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "counter(counter-name) \\"potato\\" / attr(foo) open-quote"
[FAIL] e.style['content'] = "counters(counter-name, \\".\\") \\"potato\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "counters(counter-name, \\".\\") \\"potato\\" / attr(foo) open-quote"
[FAIL] e.style['content'] = "\\"(\\" counters(counter-name, \\".\\", counter-style) \\")\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "\\"(\\" counters(counter-name, \\".\\", counter-style) \\")\\" / attr(foo) open-quote"
[FAIL] e.style['content'] = "open-quote \\"hello\\" \\"world\\" close-quote / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "open-quote \\"hello\\" \\"world\\" close-quote / attr(foo) open-quote"
[FAIL] e.style['content'] = "url(\\"https://www.example.com/picture.svg\\") \\"hello\\" / attr(foo) open-quote" should not set the property value
  assert_equals: expected "" but got "url(\\"https://www.example.com/picture.svg\\") \\"hello\\" / attr(foo) open-quote"
Harness: the test ran to completion.
